requirements.txt
# This file may be used to create an environment using:
# $ conda create --name env_name -c bioconda  --file <this file>

# REQUIRED: DO NOT CHANGE
python>=3.4.3
graphviz>=2.38.0
snakemake>=3.4.2

# Packages specific to your pipeline
ipython
seaborn
pandas
numexpr
numpy
xlrd
xlwt
blast
hmmer
pyfaidx
biopython
docopt
